BootStrap API onto MasterDetail Views

+++ src/main/java/org/infinite/spoty/forms/SaleDetailFormController.java
@@ -124,7 +124,7 @@ public class SaleDetailFormController implements Initializable {
                             SpotyThreader.spotyThreadPool(
                                     () -> {
                                         try {
-                                            SaleDetailViewModel.updateSaleDetail(getTempId());
+                                            SaleDetailViewModel.updateSaleDetail((long) getTempId());
                                         } catch (Exception e) {
                                             SpotyLogger.writeToFile(e, this.getClass());
                                         }
